feat: abstract useQuoteEstimatedFeesQuery logic outside of lending
