+++ src/pages/Lending/hooks/useGetEstimatedFeesQuery.ts
@@ -1,13 +1,12 @@
 import { useQuery } from '@tanstack/react-query'
 import { useMemo } from 'react'
+import type { EstimatedFeesQueryKey } from 'react-queries/hooks/useQuoteEstimatedFeesQuery'
 import type { EstimateFeesInput } from 'components/Modals/Send/utils'
 import { estimateFees } from 'components/Modals/Send/utils'
 import { bn } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import { selectAssetById, selectMarketDataById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
-
-import type { EstimatedFeesQueryKey } from './useQuoteEstimatedFees/types'
 
 // For use outside of react with queryClient.fetchQuery()
 export const queryFn = async ({ queryKey }: { queryKey: EstimatedFeesQueryKey }) => {
