+++ src/pages/ThorChainLP/components/AddLiquitity/AddLiquidityConfirm.tsx
@@ -1,13 +1,13 @@
 import { thorchainAssetId } from '@shapeshiftoss/caip'
 import { useCallback } from 'react'
 import { useHistory } from 'react-router'
-import type { ConfirmedQuote } from 'lib/utils/thorchain/lp/types'
+import type { LpConfirmedDepositQuote } from 'lib/utils/thorchain/lp/types'
 
 import { ReusableLpConfirm } from '../ReusableLpConfirm'
 import { AddLiquidityRoutePaths } from './types'
 
 type AddLiquidityConfirmProps = {
-  confirmedQuote: ConfirmedQuote
+  confirmedQuote: LpConfirmedDepositQuote
 }
 
 export const AddLiquidityConfirm = ({ confirmedQuote }: AddLiquidityConfirmProps) => {
