+++ src/pages/ThorChainLP/components/ReusableLpStatus/ReusableLpStatus.tsx
@@ -22,7 +22,7 @@ import { Amount } from 'components/Amount/Amount'
 import { SlideTransition } from 'components/SlideTransition'
 import { RawText } from 'components/Text'
 import { assertUnreachable } from 'lib/utils'
-import { AsymSide, type ConfirmedQuote } from 'lib/utils/thorchain/lp/types'
+import { AsymSide, type LpConfirmedDepositQuote } from 'lib/utils/thorchain/lp/types'
 import { usePools } from 'pages/ThorChainLP/queries/hooks/usePools'
 import { selectAssetById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
@@ -32,7 +32,7 @@ import { TransactionRow } from './TransactionRow'
 type ReusableLpStatusProps = {
   handleBack: () => void
   baseAssetId: AssetId
-  confirmedQuote: ConfirmedQuote
+  confirmedQuote: LpConfirmedDepositQuote
 } & PropsWithChildren
 
 export const ReusableLpStatus: React.FC<ReusableLpStatusProps> = ({
