+++ src/pages/ThorChainLP/components/ReusableLpStatus/TransactionRow.tsx
@@ -43,7 +43,7 @@ import {
 } from 'lib/utils/evm'
 import { getThorchainFromAddress, waitForThorchainUpdate } from 'lib/utils/thorchain'
 import { THORCHAIN_POOL_MODULE_ADDRESS } from 'lib/utils/thorchain/constants'
-import type { AsymSide, ConfirmedQuote } from 'lib/utils/thorchain/lp/types'
+import type { AsymSide, LpConfirmedDepositQuote } from 'lib/utils/thorchain/lp/types'
 import { depositWithExpiry } from 'lib/utils/thorchain/routerCalldata'
 import { getThorchainLpPosition } from 'pages/ThorChainLP/queries/queries'
 import { isUtxoChainId } from 'state/slices/portfolioSlice/utils'
@@ -65,7 +65,7 @@ type TransactionRowProps = {
   onComplete: () => void
   isActive?: boolean
   isLast?: boolean
-  confirmedQuote: ConfirmedQuote
+  confirmedQuote: LpConfirmedDepositQuote
   asymSide?: AsymSide | null
 }
